test(auth): decompose actor and verifier proofs with observable contention
