+++ backend/scripts/test_lane_catalogue.py
@@ -99,7 +99,25 @@ SHARED_FOUNDATION_MODULES = (
     "tests/test_submission_change_gate.py",
     "tests/test_submission_manifest.py",
     "tests/test_test_lane_evidence.py",
-    "tests/test_actors.py",
+    "tests/authentication/test_local_verifiers.py",
+    "tests/authentication/test_token_contract.py",
+    "tests/authentication/test_jwks_cache.py",
+    "tests/authentication/test_jwks_concurrency.py",
+    "tests/authentication/test_input_contracts.py",
+    "tests/authentication/test_subject_authority.py",
+    "tests/authentication/test_jwks_boundaries.py",
+    "tests/authentication/test_introspection.py",
+    "tests/authentication/test_admission.py",
+    "tests/authentication/test_development_contract.py",
+    "tests/actors/test_repository_contract.py",
+    "tests/actors/test_admin_views.py",
+    "tests/actors/test_resolution_service.py",
+    "tests/actors/test_authorization_locks.py",
+    "tests/actors/test_first_access_postgresql.py",
+    "tests/actors/test_self_api.py",
+    "tests/actors/test_self_api_lifecycle.py",
+    "tests/actors/test_identity_bounds_and_rate_controls.py",
+    "tests/actors/test_legacy_eligibility_postgresql.py",
     "tests/actors/test_compensation_adapter_eligibility.py",
     "tests/test_api_rate_controls.py",
     "tests/test_audit.py",
