+++ backend/tests/actors/conftest.py
@@ -1,0 +1,6 @@
+"""Opt-in canonical actor database/client fixtures."""
+
+from tests.actors.fixtures import (
+    actor_database_env as actor_database_env,
+    actor_client as actor_client,
+)
