+++ backend/tests/authentication/__init__.py
@@ -1,0 +1,1 @@
+"""Authentication adapter and admission proof, separate from AUTH grants."""
