+++ backend/tests/authentication/conftest.py
@@ -1,0 +1,6 @@
+"""Local cache isolation and signing fixture; no database autouse fixture."""
+
+from tests.authentication.fixtures import (
+    clear_settings_cache as clear_settings_cache,
+    rsa_signing_material as rsa_signing_material,
+)
